create plasmapy.formulary.speeds and move Alfven_speed into it
