+++ plasmapy/formulary/parameters.py
@@ -1,7 +1,6 @@
 """Functions to calculate fundamental plasma parameters."""
 
 __all__ = [
-    "Alfven_speed",
     "Bohm_diffusion",
     "ion_sound_speed",
     "kappa_thermal_speed",
@@ -19,7 +18,6 @@ __aliases__ = [
     "pth_",
     "rho_",
     "ub_",
-    "va_",
     "vth_",
     "vth_kappa_",
 ]
@@ -36,7 +34,6 @@ from typing import Optional, Union
 
 from plasmapy import particles
 from plasmapy.particles import Particle
-from plasmapy.particles.exceptions import ChargeError
 from plasmapy.utils import PhysicsError
 from plasmapy.utils.decorators import (
     bind_lite_func,
@@ -47,11 +44,25 @@ from plasmapy.utils.decorators import (
 )
 from plasmapy.utils.exceptions import PhysicsWarning, PlasmaPyFutureWarning
 
-from plasmapy.formulary import dimensionless, frequencies, lengths  # noqa
+from plasmapy.formulary import dimensionless, frequencies, lengths, speeds  # noqa
 
-__aliases__ += frequencies.__aliases__ + lengths.__aliases__
-__lite_funcs__ += frequencies.__lite_funcs__
-__all__ += frequencies.__all__ + lengths.__all__ + __aliases__ + __lite_funcs__
+__aliases__ += (
+    dimensionless.__aliases__
+    + frequencies.__aliases__
+    + lengths.__aliases__
+    + speeds.__aliases__
+)
+
+__lite_funcs__ += frequencies.__lite_funcs__ + speeds.__lite_funcs__
+
+__all__ += (
+    dimensionless.__all__
+    + frequencies.__all__
+    + lengths.__all__
+    + speeds.__all__
+    + __aliases__
+    + __lite_funcs__
+)
 
 e_si_unitless = e.value
 eps0_si_unitless = eps0.value
@@ -79,6 +90,8 @@ funcs_to_deprecate_wrap = [  # (module_name, func_name)
     ("lengths", "rhoc_"),
     ("lengths", "inertial_length"),
     ("lengths", "cwp_"),
+    ("speeds", "Alfven_speed"),
+    ("speeds", "va_"),
 ]
 for modname, name in funcs_to_deprecate_wrap:
     globals()[name] = deprecated(
@@ -229,150 +242,6 @@ def mass_density(
 
 rho_ = mass_density
 """Alias to `~plasmapy.formulary.parameters.mass_density`."""
-
-
-@check_relativistic
-@validate_quantities(density={"can_be_negative": False})
-def Alfven_speed(
-    B: u.T,
-    density: (u.m ** -3, u.kg / u.m ** 3),
-    ion: Optional[Particle] = None,
-    z_mean: Optional[numbers.Real] = None,
-) -> u.m / u.s:
-    r"""
-    Calculate the Alfvén speed.
-
-    The Alfvén speed :math:`V_A` is the typical propagation speed of magnetic
-    disturbances in a plasma, and is given by:
-
-    .. math::
-
-        V_A = \frac{B}{\sqrt{μ_0 ρ}}
-
-    where :math:`B` is the magnetic field and :math:`ρ = n_i m_i + n_e m_e`
-    is the total mass density (:math:`n_i` is the ion number density,
-    :math:`n_e` is the electron number density, :math:`m_i` is the ion mass,
-    and :math:`m_e` is the electron mass) :cite:p:`alfven:1942`.
-
-    **Aliases:** `va_`
-
-    Parameters
-    ----------
-    B : `~astropy.units.Quantity`
-        The magnetic field magnitude in units convertible to tesla.
-
-    density : `~astropy.units.Quantity`
-        Either the ion number density :math:`n_i` in units convertible to
-        m\ :sup:`-3` or the total mass density :math:`ρ` in units
-        convertible to kg m\ :sup:`-3`\ .
-
-    ion : `~plasmapy.particles.particle_class.Particle`, optional
-        Representation of the ion species (e.g., `'p'` for protons, `'D+'` for
-        deuterium, `'He-4 +1'` for singly ionized helium-4, etc.). If no charge
-        state information is provided, then the ions are assumed to be singly
-        ionized. If the density is an ion number density, then this paramter
-        is required in order to convert to mass density.
-
-    z_mean : `~numbers.Real`, optional
-        The average ionization state (arithmetic mean) of the ``ion`` composing
-        the plasma.  This is used in calculating the mass density
-        :math:`ρ = n_i (m_i + Z_{mean} m_e)`.  ``z_mean`` is ignored if
-        ``density`` is passed as a mass density and overrides any charge state
-        info provided by ``ion``.
-
-    Returns
-    -------
-    V_A : `~astropy.units.Quantity`
-        The Alfvén speed in units of m s\ :sup:`-1`.
-
-    Raises
-    ------
-    `~plasmapy.utils.exceptions.RelativityError`
-        If the Alfvén velocity is greater than or equal to the speed of light.
-
-    `TypeError`
-        If ``B`` and/or ``density`` are not of type `~astropy.units.Quantity`,
-        or convertible.
-
-    `TypeError`
-        If ``ion`` is not of type or convertible to `~plasmapy.particles.particle_class.Particle`.
-
-    `TypeError`
-        If ``z_mean`` is not of type `int` or `float`.
-
-    `~astropy.units.UnitTypeError`
-        If the magnetic field ``B`` does not have units equivalent to
-        tesla.
-
-    `~astropy.units.UnitTypeError`
-        If the ``density`` does not have units equivalent to a number density
-        or mass density.
-
-    `ValueError`
-        If ``density`` is negative.
-
-    Warns
-    -----
-    : `~plasmapy.utils.exceptions.RelativityWarning`
-        If the Alfvén velocity exceeds 5% of the speed of light.
-
-    : `~astropy.units.UnitsWarning`
-        If units are not provided for the magnetic field ``B``, units of
-        tesla are assumed.
-
-    Notes
-    -----
-    This expression does not account for relativistic effects, and
-    loses validity when the resulting speed is a significant fraction
-    of the speed of light.
-
-    Examples
-    --------
-    >>> from astropy import units as u
-    >>> from astropy.constants.si import m_p, m_e
-    >>> B = 0.014*u.T
-    >>> n = 5e19*u.m**-3
-    >>> rho = n*(m_p+m_e)
-    >>> ion = 'p'
-    >>> Alfven_speed(B, n, ion=ion)
-    <Quantity 43173.870... m / s>
-    >>> Alfven_speed(B, rho)
-    <Quantity 43173.870... m / s>
-    >>> Alfven_speed(B, rho).to(u.cm/u.us)
-    <Quantity 4.317387 cm / us>
-    >>> Alfven_speed(B, n, ion="He +2")
-    <Quantity 21664.18... m / s>
-    >>> Alfven_speed(B, n, ion="He++")
-    <Quantity 21664.18... m / s>
-    >>> Alfven_speed(B, n, ion="He", z_mean=1.8)
-    <Quantity 21661.51... m / s>
-    """
-    if density.unit.is_equivalent(u.kg / u.m ** 3):
-        rho = density
-    else:
-        if not isinstance(ion, Particle):
-            try:
-                ion = Particle(ion)
-            except TypeError:
-                raise TypeError(
-                    f"If passing a number density, you must pass a plasmapy Particle "
-                    f"(not type {type(ion)}) to calculate the mass density!"
-                )
-        if z_mean is None:
-            try:
-                z_mean = abs(ion.charge_number)
-            except ChargeError:
-                z_mean = 1
-
-        z_mean = abs(z_mean)
-        rho = mass_density(density, ion) + mass_density(density, "e", z_ratio=z_mean)
-
-    V_A = np.abs(B) / np.sqrt(mu0 * rho)
-    return V_A
-
-
-va_ = Alfven_speed
-"""Alias to `~plasmapy.formulary.parameters.Alfven_speed`."""
 
 
 @check_relativistic
